+++ development/tools/pkg/jobguard/filter.go
@@ -1,4 +1,4 @@
-package jobwaiter
+package jobguard
 
 import "strings"
