+++ development/tools/pkg/jobguard/httpclient.go
@@ -1,4 +1,4 @@
-package jobwaiter
+package jobguard
 
 import (
 	"fmt"
