+++ development/tools/pkg/jobguard/status.go
@@ -1,4 +1,4 @@
-package jobwaiter
+package jobguard
 
 import (
 	"encoding/json"
