+++ development/tools/pkg/jobguard/waiter.go
@@ -1,4 +1,4 @@
-package jobwaiter
+package jobguard
 
 import (
 	"fmt"
